requirements.txt
asgiref==3.8.1
attrs==24.2.0
coverage==7.6.1
Django==5.0.9
django-filter==24.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
Faker==30.0.0
flake8==7.1.1
inflection==0.5.1
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
pillow==10.4.0
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
uritemplate==4.1.1
